chequeo de errores

+++ apps/api/controllers/ApiCategoriesController.php
@@ -1,7 +1,8 @@
 <?php
 require_once './apps/api/controllers/ApiController.php';
 require_once './apps/models/model.php';
-require_once './apps/models/CategorieModel.php';
+require_once './apps/models/BooksModel.php';
+require_once './apps/models/CategoriesModel.php';
 
 
 
@@ -15,17 +16,17 @@ class ApiCategoriesController extends ApiController{
         $this->model = new CategoriesModel();
     }
     
-    public function get($params = [])
+    public function getCategories($params = [])
     {
         if (empty($params)) {
             $categories = $this->model->getCategories();
-            $this->view->response($categories, 200); // Le paso el JSON y el status
+            $this->view->response(['msg' => 'Datos de las categorias obtenidos con éxito', 'categories' => $categories], 200);
         } else {
             // Cambio CategoriesModel a CategorieModel (peguntar si esta bien instanciarla  asi)
-            $booksModel = new CategorieModel();
+            $booksModel = new BooksModel();
             $book = $booksModel->getBooksByCategorie($params[':ID']);
             if (!empty($book)) {
-                $this->view->response($book, 200);
+                $this->view->response(['msg' => 'Datos del los libros por categoria obtenidos con éxito', 'book' => $book], 200);
             } else {
                 $this->view->response(['msg' => "El ID ".$params[':ID'].": no existe"], 404);
                 return;
@@ -36,13 +37,18 @@ class ApiCategoriesController extends ApiController{
     public function delete($params = [])
     {
         $idCategorie = $params[':ID'];
+        
+        if (!$this->model->categoriaExiste($idCategorie)) {
+            $this->view->response(['msg' => 'La categoría ' . $idCategorie .  ' especificada no existe'], 401);
+            return;
+        }
         if($idCategorie){
             $this->model->deleteCategoria($idCategorie );
             // Eliminación exitosa, redirige a la página de categorías
             $this->view->response(['msg' => "Se elimino correctamente ".$idCategorie], 200);
         } else {
 
-            $this->view->response(['msg' => "El ID ".$idCategorie.": no existe"], 404);
+            $this->view->response(['msg' => "El ID ".$idCategorie.": no existe"], 401);
             return;
         }
     }
@@ -55,10 +61,10 @@ class ApiCategoriesController extends ApiController{
         $categorie = $body->categorie;
         //despues hacer control
         if (empty($categorie)) {
-            $this->view->response(['msg' => "Campo incompleto"], 400);
+            $this->view->response(['msg' => 'Campo incompleto'], 401);
         } else {
-            $id = $this->model->insertCategorie($categorie);
-            $this->view->response(['msg' => 'Categoría insertada con éxito'], 201);
+            $this->model->insertCategorie($categorie);
+            $this->view->response(['msg' => 'La categoría fue agregada con éxito.', 'Categoria' => $categorie], 201);
 
         }
     }
@@ -68,14 +74,18 @@ class ApiCategoriesController extends ApiController{
         $id= $params[':ID'];
         $categorie = $this->model->getCategorieById($id);
 
+        if (!$this->model->categoriaExiste($id)) {
+            $this->view->response(['msg' => 'La categoría ' . $id .  ' especificada no existe'], 401);
+            return;
+        }
         if(!empty($id)){
             if($categorie){
                 $body = $this->getData();
                 $newCategoria= $body->categorie;
                 $this->model->modifyCategorie($newCategoria, $id);
-                $this->view->response(['msg' => 'Categoría editada con éxito'], 201);
+                $this->view->response(['msg' => 'La categoría fue modificada con éxito.', 'Categoria' => $categorie], 200);
             } else {
-                $this->view->response(['msg' => "El ID ".$categorie.": no existe"], 404);
+                $this->view->response(['msg' => "El ID ".$id.": no existe"], 401);
                 return;
             }
         } else {
